# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "회원가입",
        "tags": [
          "인증 (Auth)"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "로그인 및 JWT 토큰 발급",
        "tags": [
          "인증 (Auth)"
        ]
      }
    },
    "/post": {
      "post": {
        "operationId": "PostController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시글 작성",
        "tags": [
          "게시글 (Post)"
        ]
      },
      "get": {
        "operationId": "PostController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "전체 게시글 조회",
        "tags": [
          "게시글 (Post)"
        ]
      }
    },
    "/post/{id}": {
      "get": {
        "operationId": "PostController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "단일 게시글 조회",
        "tags": [
          "게시글 (Post)"
        ]
      },
      "patch": {
        "operationId": "PostController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시글 수정",
        "tags": [
          "게시글 (Post)"
        ]
      },
      "delete": {
        "operationId": "PostController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시글 삭제",
        "tags": [
          "게시글 (Post)"
        ]
      }
    }
  },
  "info": {
    "title": "API 명세서",
    "description": "게시판 API OpenAPI",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "infoteam@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "게시글의 제목",
            "example": "옛날부터 당신같은 남자를 기다려왔다우"
          },
          "content": {
            "type": "string",
            "description": "게시글의 본문",
            "example": "정말루?"
          },
          "published": {
            "type": "boolean",
            "description": "게시 여부",
            "example": false
          },
          "categoryIds": {
            "description": "게시물 카테고리의 고유 ID 배열",
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "newCategoryNames": {
            "description": "새롭게 추가하는 카테고리의 이름 배열",
            "example": [
              "나의순결한마음을",
              "짓밟다니"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}